HEEDLS-563 Shuffle some DateHelper methods around
