+++ DigitalLearningSolutions.Data/Helpers/DateHelper.cs
@@ -7,7 +7,6 @@
 
     public static class DateHelper
     {
-        public static string StandardDateFormat = "dd/MM/yyyy";
         public static DateTime ReferenceDate => new DateTime(1905, 1, 1);
 
         public static IEnumerable<DateTime> GetPeriodsBetweenDates(
@@ -109,30 +108,6 @@
             return yearEnumerable.Select(
                 y => new DateTime(y, 1, 1)
             );
-        }
-
-        public static string GetFormatStringForGraphLabel(ReportInterval interval)
-        {
-            return interval switch
-            {
-                ReportInterval.Days => "d/M/y",
-                ReportInterval.Weeks => "wc d/M/y",
-                ReportInterval.Months => "MMM yyyy",
-                ReportInterval.Quarters => "yyyy q",
-                _ => "yyyy"
-            };
-        }
-
-        public static string GetFormatStringForDateInTable(ReportInterval interval)
-        {
-            return interval switch
-            {
-                ReportInterval.Days => "d/MM/yyyy",
-                ReportInterval.Weeks => "Week commencing d/MM/yyyy",
-                ReportInterval.Months => "MMMM, yyyy",
-                ReportInterval.Quarters => "Q, yyyy",
-                _ => "yyyy"
-            };
         }
 
         private static int ConvertMonthToQuarter(int month)
